move comment & post to PrimarySO

+++ app/viewmodel/FriendWidgetParentVM.java
@@ -3,10 +3,10 @@ package viewmodel;
 import java.util.List;
 
 public class FriendWidgetParentVM {
-	public int sn;
+	public Long sn;
 	public List<FriendWidgetChildVM> fvm;
 	
-	public FriendWidgetParentVM(int sn, List<FriendWidgetChildVM> fvm) {
+	public FriendWidgetParentVM(Long sn, List<FriendWidgetChildVM> fvm) {
 		this.sn = sn; 
 		this.fvm = fvm;
 	}
